Revert "Popravljeno racunanje PDV kod razlicitih pdv-a artikla, popravljen error prilikom otvaranja racuna u novoj godini"
This reverts commit 26a3f603c38f5963d5a2dc8692d8e1c8bcfc3cfc.

+++ Poslovni/Klase/Artikl.cs
@@ -11,7 +11,7 @@ namespace Poslovni
 {
     public struct Artikl
     {
-        public long sifra;
+        public int sifra;
         public string naziv; 
         public int kolicina;
         public int poreza_grupa;
